Renamed the ComponentRepository to ComponentFactory.
